+++ app/models/landable/traffic/user_agent_type.rb
@@ -3,7 +3,7 @@ module Landable
     class UserAgentType < ActiveRecord::Base
       include Landable::TableName
 
-      lookup_by :user_agent_type, cache: true
+      lookup_by :user_agent_type, cache: 5, find_or_create: true
 
       has_many :user_agents
     end