+++ spec/dummy/config/environments/test.rb
@@ -13,7 +13,14 @@ Dummy::Application.configure do
   config.eager_load = false
 
   # Configure static asset server for tests with Cache-Control for performance.
-  config.serve_static_assets  = true
+  # deal with Rails 4.2 deprecation of ..._assets
+  if config.respond_to?("serve_static_files")
+    config.serve_static_files = true
+  else
+    config.serve_static_assets  = true
+  end
+
+
   config.static_cache_control = "public, max-age=3600"
 
   # Show full error reports and disable caching.